+++ R/cetcolor.R
@@ -4,19 +4,59 @@
 # package load
 data(cet_color_maps, package = "cetcolor", envir = parent.env(environment()))
 
-determine_group = function(name)  color_formal_names[grepl(name, color_formal_names)]
+determine_group = function(name) {
+
+  if(name != "colorblind") {
+    search_pattern = paste0("^[", substr(name, 1, 1), "][[:digit:]]+[[:alpha:]]?$")
+  } else {
+    search_pattern = paste0("cb..[[:digit:]]+$")
+  }
+
+  color_formal_names[grepl(paste0("^.*", name,"[^_]*$"), color_formal_names) |
+                     grepl(search_pattern, color_formal_names)]
+}
 
 # For the nickname search see the cet_color_maps file for the mapping
 color_nicknames = names(alt_color_pairs)
 
 color_formal_names = names(cet_color_maps)
 
-color_attribute_names = c("rainbow","linear","isoluminant","diverging","cyclic")
+color_old_names = names(old_to_new_names)
+
+color_deprecated_names = deprecated_palettes
+
+color_attribute_names = c("rainbow", "linear", "isoluminant", "diverging", "cyclic", "colorblind")
 
 color_groups = sapply(color_attribute_names, FUN = determine_group)
 
 # Search for proper scheme name
 search_palettes = function(name) {
+
+  # Issue a warning if a deprecated palette is detected.
+  if(name %in% deprecated_palettes) {
+    warning(strwrap(
+      paste0("The color palette named `", name, "` is no longer supported. ",
+             "This palette will be deprecated in a future release. ",
+             "Please update the palette name to a new colour palette.")
+      ),
+      call. = FALSE
+    )
+    return(name)
+  }
+
+  # Inform a shortname is now available for the color palette.
+  if(name %in% color_old_names) {
+
+    new_palette_name = old_to_new_names[[name]]
+
+    message(strwrap(
+      paste0("The color palette called `", name, "` has a new shortname of `", new_palette_name, "`. ",
+             "Please consider switching over to its shortname.")
+      )
+    )
+
+    return(new_palette_name)
+  }
 
   # Search the proper name
   if(name %in% color_formal_names){ return(name) }
@@ -62,22 +102,22 @@ cet_make_scheme = function(n, name, alpha) {
 
 #' CET Perceptually Uniform Color Maps
 #'
-#' Extract \code{n} RGB Hexadecimal colors from the perceptually uniform color
-#' maps developed by \href{http://peterkovesi.com/projects/colourmaps/}{Peter Kovesi}.
+#' Extract `n` RGB Hexadecimal colors from the perceptually uniform color
+#' maps developed by [Peter Kovesi](http://peterkovesi.com/projects/colourmaps/).
 #' @export
 #'
-#' @param n      A \code{numeric} value greater than one indicating how
+#' @param n      A `numeric` value greater than one indicating how
 #'               many colors to use from the color map.
 #'
-#' @param name   A \code{string} indicating the color map to use. There
-#'               are 51 options available. Please see \code{\link{cet_color_maps}}
-#'               for more information. By default, the \code{"rainbow"} color
+#' @param name   A `string` indicating the color map to use. There
+#'               are 51 options available. Please see [cet_color_maps()]
+#'               for more information. By default, the `"rainbow"` color
 #'               scheme is used.
 #'
-#' @param alpha  A \code{numeric} value between [0, 1] that indicates the level
+#' @param alpha  A `numeric` value between \eqn{[0, 1]} that indicates the level
 #'               of transparency.
 #'
-#' @return A \code{character} vector containing the RGB hexadecimal
+#' @return A `character` vector containing the RGB hexadecimal
 #' representation of the requested color map.
 #'
 #' @examples
@@ -91,7 +131,7 @@ cet_make_scheme = function(n, name, alpha) {
 #'
 #' @references
 #' Peter Kovesi. Good Colour Maps: How to Design Them.
-#' \href{https://arxiv.org/abs/1509.03700}{arXiv:1509.03700 [cs.GR] 2015}
+#' [arXiv:1509.03700 cs.GR 2015](https://arxiv.org/abs/1509.03700)
 cet_pal = function(n, name = "rainbow", alpha = 1) {
 
   name = search_palettes(name)