+++ R/display_cet.R
@@ -1,13 +1,14 @@
 # Used for n > 1
 set_display_dim = function(num) {
-  switch(as.character(num),  #(r,c)
-         "5"  = c(3,2),
-         "6"  = c(3,2),
-         "8"  = c(4,2),
-         "14" = c(5,3),
-         "23" = c(8,3),
-         "51" = c(17,3)
-  )
+
+  is_odd_num = (num %% 2 != 0)
+
+  # Count number of rows vs. columns
+  if(num < 10) {
+    c(num %/% 2 + is_odd_num, 2)
+  } else {
+    c(num %/% 4 + is_odd_num, 4)
+  }
 }
 
 # Short function to make an image
@@ -43,10 +44,10 @@ display_cet_pal = function(n = 256, name = "rainbow", alpha = 1) {
   display_gradient(color_map, name)
 }
 
-#' @param attribute A \code{character} string indicating the attribute.
-#'                  Accepted values are: \code{"rainbow"} (Default),
-#'                  \code{"linear"}, \code{"diverging"}, \code{"cyclic"}, and
-#'                  \code{"isoluminant"}.
+#' @param attribute A `character` string indicating the attribute.
+#'                  Accepted values are: `"rainbow"` (Default),
+#'                  `"linear"`, `"diverging"`, `"cyclic"`
+#'                  `"isoluminant"`, and `"colorblind"`.
 #' @rdname display_cet
 #' @export
 display_cet_attribute = function(n = 256, attribute = "rainbow", alpha = 1) {
@@ -61,5 +62,5 @@ display_cet_attribute = function(n = 256, attribute = "rainbow", alpha = 1) {
 #' @rdname display_cet
 #' @export
 display_cet_all = function(n = 256, alpha = 1) {
-  make_gradient_grid(n, color_formal_names, alpha)
+  make_gradient_grid(n, color_formal_names[! color_formal_names %in% color_deprecated_names], alpha)
 }